Allow AGG images bigger than 2 GB (#4421)

+++ renderers/agg/include/agg_rendering_buffer.h
@@ -62,7 +62,7 @@ namespace mapserver
 			m_stride = stride;
 			if(stride < 0) 
             { 
-				m_start = m_buf - int(height - 1) * stride;
+				m_start = m_buf - (size_t)(height - 1) * stride;
 			}
         }
 
@@ -80,10 +80,10 @@ namespace mapserver
         //--------------------------------------------------------------------
 		AGG_INLINE       T* row_ptr(int, int y, unsigned) 
         { 
-            return m_start + y * m_stride; 
+            return m_start + (size_t)y * m_stride; 
         }
-		AGG_INLINE       T* row_ptr(int y)       { return m_start + y * m_stride; }
-		AGG_INLINE const T* row_ptr(int y) const { return m_start + y * m_stride; }
+		AGG_INLINE       T* row_ptr(int y)       { return m_start + (size_t)y * m_stride; }
+		AGG_INLINE const T* row_ptr(int y) const { return m_start + (size_t)y * m_stride; }
 		AGG_INLINE row_data row    (int y) const 
         { 
             return row_data(0, m_width-1, row_ptr(y));
